demo/data_to_supabase/translate_module/requirements.txt
# requirements.txt

fastapi
uvicorn
python-dotenv
google-cloud-translate